pyproject.toml
[tool.poetry]
name = "weatherdata"
version = "0.1.0"
description = ""
authors = ["atabak-hooshangi <doom_horn@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.9"
cassandra-driver = "^3.29.1"
python-dotenv = "^1.0.1"
aiohttp = "^3.9.5"
certifi = "^2024.2.2"
pytz = "^2024.1"
pandas = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
